refactor(error): move error types to dedicated modules
Move each error type to its own `error.rs` file within the respective
module for better organization.
