+++ src/patch/error.rs
@@ -1,0 +1,25 @@
+//! Error types for patch parsing.
+
+use std::borrow::Cow;
+use std::fmt;
+
+/// An error returned when parsing a `Patch` using [`Patch::from_str`] fails
+///
+/// [`Patch::from_str`]: struct.Patch.html#method.from_str
+// TODO use a custom error type instead of a Cow
+#[derive(Debug)]
+pub struct ParsePatchError(Cow<'static, str>);
+
+impl ParsePatchError {
+    pub(crate) fn new<E: Into<Cow<'static, str>>>(e: E) -> Self {
+        Self(e.into())
+    }
+}
+
+impl fmt::Display for ParsePatchError {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        write!(f, "error parsing patch: {}", self.0)
+    }
+}
+
+impl std::error::Error for ParsePatchError {}
