+++ src/patches.rs
@@ -3,16 +3,18 @@
 //! This module provides [`Patches`] for parsing patches that contain changes
 //! to multiple files, like the output of `git diff` or `git format-patch`.
 
+mod error;
 mod parse;
-pub use parse::Patches;
 #[cfg(test)]
 mod tests;
 
 use std::borrow::Cow;
-use std::fmt;
 
 use crate::binary::BinaryPatch;
 use crate::Patch;
+
+pub use error::PatchSetParseError;
+pub use parse::Patches;
 
 #[derive(Debug, Clone, Copy, Default)]
 pub(crate) enum Format {
@@ -402,21 +404,3 @@ impl FileOperation<'_> {
         matches!(self, FileOperation::Copy { .. })
     }
 }
-
-/// An error returned when parsing patches fails.
-#[derive(Debug)]
-pub struct PatchSetParseError(Cow<'static, str>);
-
-impl PatchSetParseError {
-    pub(crate) fn new<E: Into<Cow<'static, str>>>(e: E) -> Self {
-        Self(e.into())
-    }
-}
-
-impl fmt::Display for PatchSetParseError {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        write!(f, "error parsing patchset: {}", self.0)
-    }
-}
-
-impl std::error::Error for PatchSetParseError {}
